changing layout of the REST api's accoridng to new documentation
